Finalise moving files into Minimap folder
